refactor filter parsing
