+++ lib/plausible/stats/filters/dashboard_filter_parser.ex
@@ -1,63 +1,40 @@
-defmodule Plausible.Stats.Filters do
-  @visit_props [
-    "source",
-    "referrer",
-    "utm_medium",
-    "utm_source",
-    "utm_campaign",
-    "utm_content",
-    "utm_term",
-    "screen",
-    "device",
-    "browser",
-    "browser_version",
-    "os",
-    "os_version",
-    "country",
-    "region",
-    "city",
-    "entry_page",
-    "exit_page"
-  ]
+defmodule Plausible.Stats.Filters.DashboardFilterParser do
+  @moduledoc false
 
-  @event_props [
-    "name",
-    "page",
-    "goal"
-  ]
+  alias Plausible.Stats.Filters
 
-  def visit_props() do
-    @visit_props
+  @doc """
+  This function parses and prefixes the map filter format used by
+  the internal React dashboard API
+  """
+  def parse_and_prefix(filters_map) do
+    Enum.reduce(filters_map, %{}, fn {name, val}, new_filters ->
+      cond do
+        name in Filters.visit_props() ->
+          Map.put(new_filters, "visit:" <> name, filter_value(name, val))
+
+        name in Filters.event_props() ->
+          Map.put(new_filters, "event:" <> name, filter_value(name, val))
+
+        name == "props" ->
+          put_parsed_props(new_filters, name, val)
+
+        true ->
+          new_filters
+      end
+    end)
   end
 
-  def add_prefix(query) do
-    new_filters =
-      Enum.reduce(query.filters, %{}, fn {name, val}, new_filters ->
-        cond do
-          name in @visit_props ->
-            Map.put(new_filters, "visit:" <> name, filter_value(name, val))
-
-          name in @event_props ->
-            Map.put(new_filters, "event:" <> name, filter_value(name, val))
-
-          name == "props" ->
-            Enum.reduce(val, new_filters, fn {prop_key, prop_val}, new_filters ->
-              Map.put(new_filters, "event:props:" <> prop_key, filter_value(name, prop_val))
-            end)
-
-          true ->
-            new_filters
-        end
-      end)
-
-    %Plausible.Stats.Query{query | filters: new_filters}
+  defp put_parsed_props(new_filters, name, val) do
+    Enum.reduce(val, new_filters, fn {prop_key, prop_val}, new_filters ->
+      Map.put(new_filters, "event:props:" <> prop_key, filter_value(name, prop_val))
+    end)
   end
 
-  @non_escaped_pipe_regex ~r/(?<!\\)\|/
   defp filter_value(key, val) do
     {is_negated, val} = parse_negated_prefix(val)
     {is_contains, val} = parse_contains_prefix(val)
-    is_list = Regex.match?(@non_escaped_pipe_regex, val)
+    is_list = Regex.match?(Filters.non_escaped_pipe_regex(), val)
     is_wildcard = String.contains?(key, ["page", "goal"]) && String.match?(val, ~r/\*/)
     val = if is_list, do: parse_member_list(val), else: remove_escape_chars(val)
     val = if key == "goal", do: wrap_goal_value(val), else: val
@@ -87,7 +64,7 @@ defmodule Plausible.Stats.Filters do
 
   defp parse_member_list(raw_value) do
     raw_value
-    |> String.split(@non_escaped_pipe_regex)
+    |> String.split(Filters.non_escaped_pipe_regex())
     |> Enum.map(&remove_escape_chars/1)
   end
