+++ lib/plausible/stats/query.ex
@@ -10,7 +10,7 @@ defmodule Plausible.Stats.Query do
             include_imported: false
 
   require OpenTelemetry.Tracer, as: Tracer
-  alias Plausible.Stats.{FilterParser, Interval}
+  alias Plausible.Stats.{Filters, Interval}
 
   @type t :: %__MODULE__{}
 
@@ -167,7 +167,7 @@ defmodule Plausible.Stats.Query do
   end
 
   defp put_parsed_filters(query, params) do
-    Map.put(query, :filters, FilterParser.parse_filters(params["filters"]))
+    Map.put(query, :filters, Filters.parse(params["filters"]))
   end
 
   def put_filter(query, key, val) do
